Move tests for private routines back to `test_main.py`
