+++ tests/test_main.py
@@ -7,6 +7,7 @@ can be found in their respective separate test files.
 import builtins
 
 from hypothesis import given, strategies as st
+import numpy as np
 import pytest
 
 import iniabu
@@ -119,3 +120,23 @@ def test_unit_invalid(ini_default):
         ini_default.unit = unit
     err_msg = err_info.value.args[0]
     assert err_msg == f"Your selected unit {unit} is not a valid unit."
+
+
+# PRIVATE ROUTINES
+
+
+@given(ele=st.sampled_from(list(data.lodders09_elements.keys())))
+def test_get_all_isos(ini_default, ele):
+    """Ensure appropriate isotope list is returned for a given element."""
+    iso_list = []
+    for iso in ini_default.ele_dict[ele][1]:
+        iso_list.append(f"{ele}-{iso}")
+    assert ini_default._get_all_isos(ele) == iso_list
+
+
+@given(ele=st.sampled_from(list(data.lodders09_elements.keys())))
+def test_get_major_iso(ini_default, ele):
+    """Ensure that the correct major isotope is returned."""
+    index = np.array(ini_default.ele_dict[ele][2]).argmax()
+    maj_iso = f"{ele}-{ini_default.ele_dict[ele][1][index]}"
+    assert ini_default._get_major_iso(ele) == maj_iso
